Add support for multiple template directories.
This change fixes pony-goto-template when settings.TEMPLATE_DIRS
is a sequence containing multiple template directories. The
string returned by pony-get-setting is parsed to get a list of
directories to search for the template file.

Additionally, this change allows pony-goto-template to work even if
point is not at the beginning of the line containing the template
name; it can now be anywhere on that line.