Don't recalculate default each time
Recalculating the default value each time causes it to be used for the
wrong situations. `sgml-mode' should calculate the value once, and if
it is decided that `pony-mode' won't indent it that value should be
used. At the time of that decision the point could be many lines
higher and a different indentation may be appropriate there.