Fix pony-rc
`dolist's RESULT does not work the way the old code assumed.  It does
not bind a variable that can be modified to to change the returned
value.  Instead it is a form whose value becomes the value of
`dolist'.

The old code did set `ret' *globally* in the *best case*.  In the
worst case it did not set `ret' at all resulting in an "variable not
defined" error when trying to evalutate `ret' in RESULT.